WIP -- HKT would have to infect the `Item` type on parallel iterators, seems like the wrong approach. In particular, the iterable over mutable slices can't work here, because it can only deal out `&'p mut T` references, but the `Item` type defined on the iterator wants `&'data mut T` references.
The root case of the problem is that we are not able to take ownership
of the `&mut` references. In fact, the original signature was better in
that sense, since it took ownership of the iterator -- conveying that it
would never be used again. The primary problem then is that the lifetime
of the data we own is then confined to the `into_producer` function, so
we can't return it. (`&move` might actually help here; worth pondering.)